+++ packages/terra-progress-bar/examples/test-examples/ProgressBarDefault.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ProgressBar from '../../../lib/ProgressBar';
+import ProgressBar from '../../lib/ProgressBar';
 
 const ProgressBarDefault = () => (
   <ProgressBar id="progressbar" value={15} color="yellow" />